version change. updated entities, dtos and relations. improved account, user, file category and fundraiser controller and services

+++ src/main/java/com/keep/changes/config/AppConstants.java
@@ -6,6 +6,6 @@ public class AppConstants {
 	public static final String PAGE_SIZE = "10";
 	public static final String SORT_BY = "postId";
 	public static final String ORDER = "asc";
-	public static final Integer ADMIN_USER = 501;
-	public static final Integer NORMAL_USER = 502;
+	public static final Long ADMIN_USER = (long) 501;
+	public static final Long NORMAL_USER = (long) 502;
 }
